@@ -1,7 +1,4 @@
 import 'package:flutter/material.dart';
-import 'dart:convert';
-import 'package:http/http.dart' as http;
-
 import './product.dart';
 
 class Products with ChangeNotifier {
@@ -66,47 +63,15 @@ class Products with ChangeNotifier {
   //   notifyListeners();
   // }
 
-   Future<void> addProduct(Product product) {
-    final url =
-        Uri.https('logup-a8317-default-rtdb.firebaseio.com', 'products.json');
- return  http
-        .post(
-      url,
-      body: json.encode({
-        'tittle': product.title,
-        'description': product.description,
-        'price': product.price,
-        'imageUrl': product.imageUrl,
-        'isFavourite': product.isFavorite,
-      }),
-    )
-        .then((response) {
-      final newProduct = Product(
+  void addProduct(Product product) {
+    final newproduct = Product(
+        id: DateTime.now().toString(),
         title: product.title,
         description: product.description,
         price: product.price,
-        imageUrl: product.imageUrl,
-        id: json.decode(response.body)['name'],
-      );
-      _items.add(newProduct);
-
-      notifyListeners();
-    });
-  }
-  // _items.insert(0, newProduct); // at the start of the list
-
-  void updateProduct(String id, Product newProduct) {
-    final prodIndex = _items.indexWhere((prod) => prod.id == id);
-    if (prodIndex >= 0) {
-      _items[prodIndex] = newProduct;
-      notifyListeners();
-    } else {
-      print('...');
-    }
-  }
-
-  void deleteProduct(String id) {
-    _items.removeWhere((prod) => prod.id == id);
+        imageUrl: product.imageUrl);
+    // _items.add(value);
+    _items.add(newproduct);
     notifyListeners();
   }
 }
