@@ -1,23 +1,17 @@
+// ignore_for_file: must_be_immutable
+
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../providers/product.dart';
 import '../providers/products.dart';
 
-class EditProductScreen extends StatefulWidget {
-  // static const routeName = '/edit-product';
-  final String? Productid;
-
-  EditProductScreen({this.Productid});
-  @override
-  _EditProductScreenState createState() => _EditProductScreenState();
-}
-
-class _EditProductScreenState extends State<EditProductScreen> {
-  final _priceFocusNode = FocusNode();
-  final _descriptionFocusNode = FocusNode();
-  final _imageUrlController = TextEditingController();
-  final _imageUrlFocusNode = FocusNode();
+class EditScreenProduct extends StatefulWidget {
+  static const routeName = '/EditScreenProduct';
+  final priceFocusNode = FocusNode();
+  final descriptionFocusNode = FocusNode();
+  final urlFocusNode = FocusNode();
+  final imageurlcontroller = TextEditingController();
   final _form = GlobalKey<FormState>();
   var _editedProduct = Product(
     id: '',
@@ -26,254 +20,183 @@ class _EditProductScreenState extends State<EditProductScreen> {
     description: '',
     imageUrl: '',
   );
-  var _initValues = {
-    'title': '',
-    'description': '',
-    'price': '',
-    'imageUrl': '',
-  };
-  var _isInit = true;
 
-  var _isLoading = false;
-
-  @override
-  void initState() {
-    _imageUrlFocusNode.addListener(_updateImageUrl);
-    super.initState();
-  }
-
-  @override
-  void didChangeDependencies() {
-    if (_isInit) {
-      final productId = widget.Productid;
-      if (productId != '') {
-        _editedProduct =
-            Provider.of<Products>(context, listen: false).findById(productId!);
-        _initValues = {
-          'title': _editedProduct.title,
-          'description': _editedProduct.description,
-          'price': _editedProduct.price.toString(),
-          // 'imageUrl': _editedProduct.imageUrl,
-          'imageUrl': '',
-        };
-        _imageUrlController.text = _editedProduct.imageUrl;
-      }
-    }
-    _isInit = false;
-    super.didChangeDependencies();
-  }
-
-  @override
   void dispose() {
-    _imageUrlFocusNode.removeListener(_updateImageUrl);
-    _priceFocusNode.dispose();
-    _descriptionFocusNode.dispose();
-    _imageUrlController.dispose();
-    _imageUrlFocusNode.dispose();
-    super.dispose();
+    //...
+    descriptionFocusNode.dispose();
+    priceFocusNode.dispose();
+    imageurlcontroller.dispose();
+    // super.dispose();
+    //...
   }
 
-  void _updateImageUrl() {
-    if (!_imageUrlFocusNode.hasFocus) {
-      if ((!_imageUrlController.text.startsWith('http') &&
-              !_imageUrlController.text.startsWith('https')) ||
-          (!_imageUrlController.text.endsWith('.png') &&
-              !_imageUrlController.text.endsWith('.jpg') &&
-              !_imageUrlController.text.endsWith('.jpeg'))) {
-        return;
-      }
-      setState(() {});
-    }
-  }
+  EditScreenProduct({super.key});
 
-  void _saveForm() {
-    final isValid = _form.currentState!.validate();
-    if (!isValid) {
+  @override
+  State<EditScreenProduct> createState() => _EditScreenProductState();
+}
+
+class _EditScreenProductState extends State<EditScreenProduct> {
+  void _saveForm(BuildContext context) {
+    final isvalid = widget._form.currentState!.validate();
+    if (!isvalid) {
       return;
     }
-    _form.currentState!.save();
-    setState(() {
-      _isLoading = true;
-    });
-    if (_editedProduct.id != '') {
-      Provider.of<Products>(context, listen: false)
-          .updateProduct(_editedProduct.id, _editedProduct);
-      setState(() {
-        Navigator.of(context).pop();
-        _isLoading = false;
-      });
-    } else {
-      Provider.of<Products>(context, listen: false)
-          .addProduct(_editedProduct)
-          .then((value) => setState(() {
-                Navigator.of(context).pop();
-                _isLoading = false;
-              }));
-    }
+    widget._form.currentState?.save();
+    Provider.of<Products>(context, listen: false)
+        .addProduct(widget._editedProduct);
+    Navigator.of(context).pop();
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Edit Product'),
+        title: Text("Edit Product "),
         actions: <Widget>[
           IconButton(
             icon: Icon(Icons.save),
-            onPressed: _saveForm,
+            onPressed: () {
+              _saveForm(context);
+            },
           ),
         ],
       ),
-      body: _isLoading ? Center(child: CircularProgressIndicator(),) : Padding(
+      body: Padding(
         padding: const EdgeInsets.all(16.0),
         child: Form(
-          key: _form,
+          key: widget._form,
           child: ListView(
-            children: <Widget>[
+            children: [
               TextFormField(
-                initialValue: _initValues['title'],
-                decoration: InputDecoration(labelText: 'Title'),
+                decoration: InputDecoration(labelText: 'Tittle'),
                 textInputAction: TextInputAction.next,
-                onFieldSubmitted: (_) {
-                  FocusScope.of(context).requestFocus(_priceFocusNode);
-                },
-                validator: (value) {
-                  if (value!.isEmpty) {
-                    return 'Please provide a value.';
+                validator: (values) {
+                  if (values!.isEmpty) {
+                    return 'Please Provide A Value';
                   }
                   return null;
                 },
+                onFieldSubmitted: (_) {
+                  FocusScope.of(context).requestFocus(widget.priceFocusNode);
+                },
                 onSaved: (value) {
-                  _editedProduct = Product(
-                      title: value!,
-                      price: _editedProduct.price,
-                      description: _editedProduct.description,
-                      imageUrl: _editedProduct.imageUrl,
-                      id: _editedProduct.id,
-                      isFavorite: _editedProduct.isFavorite);
+                  widget._editedProduct = Product(
+                    title: value!,
+                    price: widget._editedProduct.price,
+                    description: widget._editedProduct.description,
+                    imageUrl: widget._editedProduct.imageUrl,
+                    id: '',
+                  );
                 },
               ),
               TextFormField(
-                initialValue: _initValues['price'],
                 decoration: InputDecoration(labelText: 'Price'),
                 textInputAction: TextInputAction.next,
                 keyboardType: TextInputType.number,
-                focusNode: _priceFocusNode,
+                focusNode: widget.priceFocusNode,
                 onFieldSubmitted: (_) {
-                  FocusScope.of(context).requestFocus(_descriptionFocusNode);
+                  FocusScope.of(context)
+                      .requestFocus(widget.descriptionFocusNode);
                 },
-                validator: (value) {
-                  if (value!.isEmpty) {
-                    return 'Please enter a price.';
+                validator: (values) {
+                  if (values!.isEmpty) {
+                    return 'Please Enter Price';
                   }
-                  if (double.tryParse(value) == null) {
-                    return 'Please enter a valid number.';
+                  if (double.tryParse(values) == null) {
+                    return 'Please Enter A Valid Number';
                   }
-                  if (double.parse(value) <= 0) {
-                    return 'Please enter a number greater than zero.';
+                  if (double.parse(values) <= 0) {
+                    return 'Enter Value Grater Then 0';
                   }
                   return null;
                 },
                 onSaved: (value) {
-                  _editedProduct = Product(
-                      title: _editedProduct.title,
-                      price: double.parse(value!),
-                      description: _editedProduct.description,
-                      imageUrl: _editedProduct.imageUrl,
-                      id: _editedProduct.id,
-                      isFavorite: _editedProduct.isFavorite);
+                  widget._editedProduct = Product(
+                    title: widget._editedProduct.title,
+                    price: double.parse(value!),
+                    description: widget._editedProduct.description,
+                    imageUrl: widget._editedProduct.imageUrl,
+                    id: '',
+                  );
                 },
               ),
               TextFormField(
-                initialValue: _initValues['description'],
                 decoration: InputDecoration(labelText: 'Description'),
-                maxLines: 3,
+                textInputAction: TextInputAction.next,
                 keyboardType: TextInputType.multiline,
-                focusNode: _descriptionFocusNode,
-                validator: (value) {
-                  if (value!.isEmpty) {
-                    return 'Please enter a description.';
+                maxLines: 3,
+                focusNode: widget.descriptionFocusNode,
+                validator: (values) {
+                  if (values!.isEmpty) {
+                    return 'Please Provide A Description';
                   }
-                  if (value.length < 10) {
-                    return 'Should be at least 10 characters long.';
+                  if (values.length < 10) {
+                    return 'MiniMum Enter 10 Char';
                   }
                   return null;
                 },
                 onSaved: (value) {
-                  _editedProduct = Product(
-                    title: _editedProduct.title,
-                    price: _editedProduct.price,
+                  widget._editedProduct = Product(
+                    title: widget._editedProduct.title,
+                    price: widget._editedProduct.price,
                     description: value!,
-                    imageUrl: _editedProduct.imageUrl,
-                    id: _editedProduct.id,
-                    isFavorite: _editedProduct.isFavorite,
+                    imageUrl: widget._editedProduct.imageUrl,
+                    id: '',
                   );
                 },
               ),
               Row(
                 crossAxisAlignment: CrossAxisAlignment.end,
-                children: <Widget>[
+                mainAxisAlignment: MainAxisAlignment.spaceAround,
+                children: [
                   Container(
-                    width: 100,
+                    margin: EdgeInsets.only(top: 20),
                     height: 100,
-                    margin: EdgeInsets.only(
-                      top: 8,
-                      right: 10,
-                    ),
-                    decoration: BoxDecoration(
-                      border: Border.all(
-                        width: 1,
-                        color: Colors.grey,
-                      ),
-                    ),
-                    child: _imageUrlController.text.isEmpty
-                        ? Text('Enter a URL')
+                    width: 100,
+                    color: Colors.grey,
+                    child: widget.imageurlcontroller.text.isEmpty
+                        ? Text("Enter Image Url")
                         : FittedBox(
-                            child: Image.network(
-                              _imageUrlController.text,
-                              fit: BoxFit.cover,
-                            ),
+                            child:
+                                Image.network(widget.imageurlcontroller.text),
+                            fit: BoxFit.cover,
                           ),
                   ),
                   Expanded(
                     child: TextFormField(
-                      decoration: InputDecoration(labelText: 'Image URL'),
-                      keyboardType: TextInputType.url,
+                      decoration: InputDecoration(labelText: 'Enter Image Url'),
                       textInputAction: TextInputAction.done,
-                      controller: _imageUrlController,
-                      focusNode: _imageUrlFocusNode,
-                      onFieldSubmitted: (_) {
-                        _saveForm();
+                      keyboardType: TextInputType.url,
+                      onFieldSubmitted: (text) {
+                        setState(() {});
                       },
-                      validator: (value) {
-                        if (value!.isEmpty) {
-                          return 'Please enter an image URL.';
+                      controller: widget.imageurlcontroller,
+                      // maxLines: 3,
+                      focusNode: widget.urlFocusNode,
+                      validator: ((values) {
+                        if (values!.isEmpty) {
+                          return 'Please Enter Image Url';
                         }
-                        if (!value.startsWith('http') &&
-                            !value.startsWith('https')) {
-                          return 'Please enter a valid URL.';
+                        if (!values.startsWith('http') &&
+                            !values.startsWith('https')) {
+                          return 'Enter A Valid Url - ex. https//';
                         }
-                        if (!value.endsWith('.png') &&
-                            !value.endsWith('.jpg') &&
-                            !value.endsWith('.jpeg')) {
-                          return 'Please enter a valid image URL.';
-                        }
-                        return null;
-                      },
+                      }),
                       onSaved: (value) {
-                        _editedProduct = Product(
-                          title: _editedProduct.title,
-                          price: _editedProduct.price,
-                          description: _editedProduct.description,
-                          imageUrl: value!,
-                          id: _editedProduct.id,
-                          isFavorite: _editedProduct.isFavorite,
-                        );
+                        widget
+                          .._editedProduct = Product(
+                            title: widget._editedProduct.title,
+                            price: widget._editedProduct.price,
+                            description: widget._editedProduct.description,
+                            imageUrl: value!,
+                            id: '',
+                          );
                       },
                     ),
                   ),
                 ],
-              ),
+              )
             ],
           ),
         ),
