@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 class Badge extends StatelessWidget {
   const Badge({
+    
     required this.child,
     required this.value,
      this.color,
@@ -25,7 +26,7 @@ class Badge extends StatelessWidget {
             // color: Theme.of(context).accentColor,
             decoration: BoxDecoration(
               borderRadius: BorderRadius.circular(10.0),
-              color: color != null ? color : Theme.of(context).colorScheme.secondary,
+              color: color != null ? color : Theme.of(context).accentColor,
             ),
             constraints: BoxConstraints(
               minWidth: 16,
