@@ -39,28 +39,24 @@ class CartItem extends StatelessWidget {
       direction: DismissDirection.endToStart,
       confirmDismiss: (direction) {
         return showDialog(
-          context: context,
-          builder: (ctx) => AlertDialog(
-                title: Text('Are you sure?'),
-                content: Text(
-                  'Do you want to remove the item from the cart?',
-                ),
-                actions: <Widget>[
-                  TextButton(
-                    child: Text('No'),
-                    onPressed: () {
-                      Navigator.of(ctx).pop(false);
-                    },
-                  ),
-                  TextButton(
-                    child: Text('Yes'),
-                    onPressed: () {
-                      Navigator.of(ctx).pop(true);
-                    },
-                  ),
-                ],
-              ),
-        );
+            context: context,
+            builder: (ctx) => AlertDialog(
+                  title: Text('Are You Sure'),
+                  content: Text("Do You Want Remove This Item From Cart?"),
+                  actions: [
+                    TextButton(
+                      onPressed: () {
+                        Navigator.pop(ctx, false);
+                      },
+                      child: Text('No'),
+                    ),
+                    TextButton(
+                        onPressed: () {
+                          Navigator.pop(ctx, true);
+                        },
+                        child: Text('Yes'))
+                  ],
+                ));
       },
       onDismissed: (direction) {
         Provider.of<Cart>(context, listen: false).removeItem(productId);
