@@ -1,22 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-
-import '../screens/edit_product_screen.dart';
-import '../providers/products.dart';
+import 'package:flutter_complete_guide/providers/product.dart';
+import 'package:flutter_complete_guide/screens/edit_product_screen.dart';
 
 class UserProductItem extends StatelessWidget {
-  final String id;
-  final String title;
-  final String imageUrl;
+  Product item;
 
-  UserProductItem(this.id, this.title, this.imageUrl);
+  UserProductItem({
+    required this.item,
+  });
 
   @override
   Widget build(BuildContext context) {
     return ListTile(
-      title: Text(title),
+      title: Text(item.title),
       leading: CircleAvatar(
-        backgroundImage: NetworkImage(imageUrl),
+        backgroundImage: NetworkImage(item.imageUrl),
       ),
       trailing: Container(
         width: 100,
@@ -25,23 +23,17 @@ class UserProductItem extends StatelessWidget {
             IconButton(
               icon: Icon(Icons.edit),
               onPressed: () {
-                // Navigator.of(context)
-                //     .pushNamed(EditProductScreen(Productid: id));
-                Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => EditProductScreen(
-                        Productid: id,
-                      ),
-                    ));
+                Navigator.of(
+                  context,
+                ).pushNamed(
+                  EditScreenProduct.routeName,
+                );
               },
               color: Theme.of(context).primaryColor,
             ),
             IconButton(
               icon: Icon(Icons.delete),
-              onPressed: () {
-                Provider.of<Products>(context, listen: false).deleteProduct(id);
-              },
+              onPressed: () {},
               color: Theme.of(context).errorColor,
             ),
           ],
